requirements.txt
# ./requirements.txt
Django==1.6.2
argparse==1.2.1
distribute==0.6.34
wsgiref==0.1.2
django-cron==0.3.3
django-widget-tweaks==1.3
django-braces==1.4.0
Pygments==1.6
